Add argument for build specific environment
It is likely that a user may want to specify environment variables
during the build phase as well as during pnpm install. Add an argument
for a build specific environment. This allows the user to avoid
duplicating the environment between a configure or prebuild step and
installEnv, by using something like:

let
  sharedEnv = { ... };
  buildEnvOverrides = { ... };
  installEnvOverrides = { ... };
in
  mkPnpmPackage {
    buildEnv = sharedEnv // buildEnvOverrides;
    installEnv = sharedEnv // installEnvOverrides;
    ...
  }