Add support for placing additional artifacts into the result.
A user may want the node_modules and package artifacts as a part of the
output in order to have a runnable build artifact without needing the
node-modules derivation or the sources available. Add options to control
additional copies as a part of the install.

Simplify the values used in the derivation so that the conditional logic
is almost all in the let rather than in the derivation body.

Some of this change should probably rebased into the first commit.